Warn if number of rows added doesn't match
We seem to be having an odd bug where some of the people with multiple
memberships in the term only end up with a single row stored in the database.

To help debug this, warn if the number of rows for a person doesn't
match the number of memberships we just saved for them.