Clean out table by DROPping it
Simply DELETE-ing all the data retains the old structure, and any unique
indexes that were set up before. In this specific case, the unique key
used to solely be `(:id, :term)`, rather than the current `(:id, :term, :
start_date)`, so the current version of the table already has that index
(which gets created at `CREATE TABLE` time, which in normal usage is on
first write to it). This has caused a very subtle bug, where we've been
losing the second membership row for people for several months, and not
noticing because local runs, which didn't already have the old table,
were getting it right.

This is something we're going to need to look at doing everywhere from
now on to avoid this class of problem.