Only scrape members of 52nd Parliament
Combining 51 and 52 doesn't actually get us members of both, as the URL
we're scraping is only current members. Instead it gets us the
membership data of the 51st term for the current members, which isn't
the right thing.

Later we'll expand the scraper to also include historic people, but for
now keep this as only scraping current members.